+++ InfiniteGUI-DLL/Notification.cpp
@@ -77,7 +77,7 @@ bool Notification::Done() const
 
 void Notification::SetMaxElementPos()
 {
-    ImVec2 maxScreenSize = ImVec2(opengl_hook::screen_size.x, opengl_hook::screen_size.y);
+    ImVec2 maxScreenSize = ImVec2((float)opengl_hook::screen_size.x, (float)opengl_hook::screen_size.y);
     startElement.pos = ImVec2(maxScreenSize.x + windowPadding, maxScreenSize.y - (placeIndex + 1.5f) * (size.y + windowPadding));
     targetElement.pos = ImVec2(maxScreenSize.x - size.x - windowPadding, startElement.pos.y);
 }
